requirements.txt
Django>=5.2.3
Pillow>=10.0.0
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.2
boto3>=1.26.0
django-storages>=1.14.0
python-dotenv>=1.0.0 